requirements.txt
transformers
torch
bitsandbytes
accelerate
pillow
numpy
pandas
python-dotenv
